03/15/2025

+++ projects/assembly-emulator/assembly.h
@@ -50,7 +50,7 @@ class Assembly
 
         struct LRegister
         {
-            int8_t int_value;
+            short int int_value;
         };
 
         // struct Stack
